+++ sphinx/directives/other.py
@@ -421,99 +421,6 @@ class Include(BaseInclude, SphinxDirective):
         return super().run()
 
 
-class Collapsible(SphinxDirective):
-    """
-    Directive to mark collapsible content, with an optional summary line.
-    """
-    has_content = True
-    optional_arguments = 1
-    final_argument_whitespace = True
-    option_spec: OptionSpec = {
-        'class': directives.class_option,
-        'open': directives.flag,
-    }
-
-    @staticmethod
-    def _prepare_argument_string(s: str) -> str:
-        """Prepare a directive argument string.
-
-        Remove common leading indentation, where the indentation of the first
-        line is ignored.
-
-        Return a list of lines usable for inserting into a docutils StringList.
-        """
-        lines = s.expandtabs().splitlines()
-
-        # Find minimum indentation of any non-blank lines after the first.
-        # if your indent is larger than a million yotta-spaces,  there's a problem...
-        margin = 10**30
-        for line in lines[1:]:
-            content = len(line.lstrip())
-            if content:
-                indent = len(line) - content
-                margin = min(margin, indent)
-
-        if margin == 10**30:
-            return s
-
-        return "\n".join((lines[0], *(line[margin:] for line in lines[1:])))
-
-    def run(self) -> List[Node]:
-        node = addnodes.collapsible(collapsible_open="open" in self.options)
-        node.document = self.state.document
-        self.set_source_info(node)
-        if len(self.arguments) > 0:
-            summary = self.arguments[0].strip()
-        else:
-            summary = "Collapsed Content:"
-
-        # parse the argument as reST
-        summary_trimmed = self._prepare_argument_string(summary)
-        textnodes, _messages = self.state.inline_text(summary_trimmed, self.lineno)
-        node.append(addnodes.collapsible_summary(summary, "", *textnodes))
-
-        # Same as util.nested_parse_with_titles but try to handle nested
-        # sections which should be raised higher up the doctree.
-        memo: Any = self.state.memo
-        surrounding_title_styles = memo.title_styles
-        surrounding_section_level = memo.section_level
-        memo.title_styles = []
-        memo.section_level = 0
-        try:
-            self.state.nested_parse(self.content, self.content_offset,
-                                    node, match_titles=True)
-            title_styles = memo.title_styles
-            if (not surrounding_title_styles or
-                    not title_styles or
-                    title_styles[0] not in surrounding_title_styles or
-                    not self.state.parent):
-                # No nested sections so no special handling needed.
-                return [node]
-            # Calculate the depths of the current and nested sections.
-            current_depth = 0
-            parent = self.state.parent
-            while parent:
-                current_depth += 1
-                parent = parent.parent
-            current_depth -= 2
-            title_style = title_styles[0]
-            nested_depth = len(surrounding_title_styles)
-            if title_style in surrounding_title_styles:
-                nested_depth = surrounding_title_styles.index(title_style)
-            # Use these depths to determine where the nested sections should
-            # be placed in the doctree.
-            n_sects_to_raise = current_depth - nested_depth + 1
-            parent = cast(nodes.Element, self.state.parent)
-            for _i in range(n_sects_to_raise):
-                if parent.parent:
-                    parent = parent.parent
-            parent.append(node)
-            return []
-        finally:
-            memo.title_styles = surrounding_title_styles
-            memo.section_level = surrounding_section_level
-
-
 def setup(app: Sphinx) -> ExtensionMetadata:
     directives.register_directive('toctree', TocTree)
     directives.register_directive('sectionauthor', Author)
@@ -526,7 +433,6 @@ def setup(app: Sphinx) -> ExtensionMetadata:
     directives.register_directive('hlist', HList)
     directives.register_directive('only', Only)
     directives.register_directive('include', Include)
-    directives.register_directive('collapsible', Collapsible)
 
     # register the standard rst class directive under a different name
     # only for backwards compatibility now