+++ sphinx/transforms/post_transforms/__init__.py
@@ -15,8 +15,7 @@ from sphinx.locale import __
 from sphinx.transforms import SphinxTransform
 from sphinx.util import logging
 from sphinx.util.docutils import SphinxTranslator
-from sphinx.util.nodes import (find_pending_xref_condition, process_collapsible_nodes,
-                               process_only_nodes)
+from sphinx.util.nodes import find_pending_xref_condition, process_only_nodes
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -233,20 +232,6 @@ class OnlyNodeTransform(SphinxPostTransform):
         process_only_nodes(self.document, self.app.builder.tags)
 
 
-class CollapsibleNodeTransform(SphinxPostTransform):
-    default_priority = 55
-
-    def run(self, **kwargs: Any) -> None:
-        # A comment on the comment() nodes being inserted: replacing by [] would
-        # result in a "Losing ids" exception if there is a target node before
-        # the only node, so we make sure docutils can transfer the id to
-        # something, even if it's just a comment and will lose the id anyway...
-        builder_name = self.app.builder.name
-        if getattr(self.app.builder.config, "html4_writer", False):
-            builder_name = "html4"
-        process_collapsible_nodes(self.document, builder_name)
-
-
 class SigElementFallbackTransform(SphinxPostTransform):
     """Fallback various desc_* nodes to inline if translator does not support them."""
 
@@ -304,7 +289,6 @@ class PropagateDescDomain(SphinxPostTransform):
 def setup(app: Sphinx) -> ExtensionMetadata:
     app.add_post_transform(ReferencesResolver)
     app.add_post_transform(OnlyNodeTransform)
-    app.add_post_transform(CollapsibleNodeTransform)
     app.add_post_transform(SigElementFallbackTransform)
     app.add_post_transform(PropagateDescDomain)