+++ CHANGES.rst
@@ -46,8 +46,9 @@ Features added
 
   Patch by Adam Turner.
 
-* #10532: Added the :rst:dir:`collapsible` directive and support for collapsible
-  content in HTML. Patch by Adam Turner.
+* #10532: Add a new extension to support collapsible content in HTML,
+  ``sphinx.ext.collapse``, which enables the :rst:dir:`collapsible` directive.
+  Patch by Adam Turner.
 
 Bugs fixed
 ----------